refactor: move common logic to a module
