+++ common/globalConfig.ts
