+++ common/globalTypes.ts
